# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "REST API",
      "description": "API documentation",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://142.93.60.109:3000",
        "description": "DigitalOcean server"
      }
    ],
    "paths": {
      "/agents": {
        "get": {
          "summary": "Get all agents",
          "responses": {
            "200": {
              "description": "List of agents",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Agent"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/companies": {
        "get": {
          "summary": "Get all companies",
          "responses": {
            "200": {
              "description": "List of companies",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Company"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/company": {
        "post": {
          "summary": "Create a new company",
          "description": "Creates a new company in the database",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Company created successfully"
            },
            "400": {
              "description": "Invalid request body or parameters"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/company/{COMPANY_ID}": {
        "put": {
          "summary": "Update entire company",
          "description": "Update the entire company object",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "COMPANY_ID",
              "description": "Id of company getting updated",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "COMPANY_NAME",
              "description": "Company to be updated",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/components/schemas/CompanyPutRequest"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Operation"
            },
            "400": {
              "description": "Bad Request"
            },
            "500": {
              "description": "Error"
            }
          }
        },
        "patch": {
          "summary": "Update company name",
          "description": "Update company name",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "COMPANY_ID",
              "description": "Id of company getting updated",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "COMPANY_NAME",
              "description": "Company to be updated",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/components/schemas/CompanyPatchRequest"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Operation"
            },
            "400": {
              "description": "Bad Request"
            },
            "500": {
              "description": "Error"
            }
          }
        },
        "delete": {
          "summary": "Delete a company",
          "description": "Deletes a company from the database",
          "parameters": [
            {
              "name": "COMPANY_ID",
              "description": "Id of company getting deleted",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Company deleted successfully"
            },
            "400": {
              "description": "Invalid company ID supplied"
            },
            "500": {
              "description": "Error"
            }
          }
        }
      },
      "/customers": {
        "get": {
          "summary": "Get all customers",
          "responses": {
            "200": {
              "description": "List of customers",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Customer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/customers/{CUST_CODE}": {
        "patch": {
          "summary": "Update customer information",
          "description": "Updates a customer's information in the database",
          "parameters": [
            {
              "name": "CUST_CODE",
              "description": "Code of customer getting updated",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "customer",
              "description": "Customer to be updated",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/components/schemas/CustomerPatchRequest"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Operation"
            },
            "400": {
              "description": "Bad Request"
            },
            "500": {
              "description": "Error"
            }
          }
        }
      },
      "/food/name/{id}": {
        "patch": {
          "summary": "Update food item name",
          "description": "Update the name of a food item identified by its ID.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the food item to be updated",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "food",
              "in": "body",
              "description": "Updated food information",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "The updated name of the food item"
                  }
                },
                "required": ["name"]
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Food item updated successfully"
            },
            "400": {
              "description": "Invalid request body or parameters"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      }    
    },
    "components": {
      "schemas": {
        "Agent": {
          "type": "object",
          "properties": {
            "AGENT_CODE": {
              "type": "string"
            },
            "AGENT_NAME": {
              "type": "string"
            },
            "WORKING_AREA": {
              "type": "string"
            },
            "COMMISSION": {
              "type": "number",
              "format": "float"
            },
            "PHONE_NO": {
              "type": "string"
            },
            "COUNTRY": {
              "type": "string"
            }
          },
          "required": [
            "AGENT_CODE"
          ]
        },
        "Company": {
          "type": "object",
          "properties": {
            "COMPANY_ID": {
              "type": "string",
              "description": "The unique identifier for the company."
            },
            "COMPANY_NAME": {
              "type": "string",
              "description": "The name of the company."
            },
            "COMPANY_CITY": {
              "type": "string",
              "description": "The city where the company is located."
            }
          },
          "required": [
            "COMPANY_ID",
            "COMPANY_NAME",
            "COMPANY_CITY"
          ],
          "description": "Schema representing a company."
        },
        "Customer": {
          "type": "object",
          "properties": {
            "CUST_CODE": {
              "type": "string",
              "description": "The unique identifier for the customer."
            },
            "CUST_NAME": {
              "type": "string",
              "description": "The name of the customer."
            },
            "CUST_CITY": {
              "type": "string",
              "description": "The city where the customer is located."
            },
            "WORKING_AREA": {
              "type": "string",
              "description": "The area where the customer operates or works."
            },
            "CUST_COUNTRY": {
              "type": "string",
              "description": "The country where the customer is located."
            },
            "GRADE": {
              "type": "number",
              "format": "float",
              "description": "The grade or rating of the customer."
            },
            "OPENING_AMT": {
              "type": "number",
              "format": "float",
              "description": "The opening amount for the customer."
            },
            "RECEIVE_AMT": {
              "type": "number",
              "format": "float",
              "description": "The amount received from the customer."
            },
            "PAYMENT_AMT": {
              "type": "number",
              "format": "float",
              "description": "The amount paid by the customer."
            },
            "OUTSTANDING_AMT": {
              "type": "number",
              "format": "float",
              "description": "The outstanding amount for the customer."
            },
            "PHONE_NO": {
              "type": "string",
              "description": "The phone number of the customer."
            },
            "AGENT_CODE": {
              "type": "string",
              "description": "The agent code associated with the customer."
            }
          },
          "required": [
            "CUST_CODE",
            "CUST_NAME",
            "CUST_CITY",
            "WORKING_AREA",
            "CUST_COUNTRY",
            "GRADE",
            "OPENING_AMT",
            "RECEIVE_AMT",
            "PAYMENT_AMT",
            "OUTSTANDING_AMT",
            "PHONE_NO",
            "AGENT_CODE"
          ],
          "description": "Schema representing a customer."
        },
        "CompanyPatchRequest": {
          "type": "object",
          "properties": {
            "COMPANY_NAME": {
              "type": "string",
              "description": "The updated name of the company."
            }
          },
          "required": [
            "COMPANY_NAME"
          ],
          "description": "Request body for updating company name."
        },
        "CompanyPutRequest": {
          "type": "object",
          "properties": {
            "COMPANY_NAME": {
              "type": "string",
              "description": "The updated name of the company."
            },
            "COMPANY_CITY": {
              "type": "string",
              "description": "The updated city where the company is located."
            }
          },
          "required": [
            "COMPANY_NAME",
            "COMPANY_CITY"
          ],
          "description": "Request body for updating a company."
        },
        "CustomerPatchRequest": {
          "type": "object",
          "properties": {
            "CUST_NAME": {
              "type": "string",
              "description": "The updated name of the customer."
            }
          },
          "required": [
            "CUST_NAME"
          ],
          "description": "Request body for updating customer name."
        }
      }
    }
  }
  